qt4: fix incorrect architecture for qt4 tools
Don't install the native staging versions of the qt4 tools in the bin
directory; we have been doing this for a very long time but it is no
longer necessary since we specify them via the QMAKE_UIC, QMAKE_MOC etc.
variables. Doing so was sometimes resulting in those executables being
the ones that end up being packaged (probably only sometimes due to
different date/time values on the staging files and the interaction
with make).

Fixes [YOCTO #1856].

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>