kern-tools: import flexible branching support
The commit:

   kern-tools: update SRCREV to pickup git operation fixes

Brought in the ability to trap failed git operations on the working tree,
but what it missed were some branching changes that allow arbitrary
branch points and the ability to create a branch multiple times (if a feature
is included multiple times). The graphics driver branches used by some
BSPs need this part of the change to properly handle graphics driver
branches.

Updating the SRCREV to pickup the associated kern-tools changes for this
support.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>