gtk-icon-cache.bbclass: add -f option to gtk-update-icon-cache in postinst
The gtk-update-icon-cache utility does not create a cache file if
one does not already exist.  Since some packages (notably gnome-icon-theme)
do not ship a cache file by default, the results of the utility are
not saved.  Adding the force option writes out the cache even if there
wasn't a default cache in the package.

Signed-off-by: Steve Sakoman <steve@sakoman.com>