autotools.bbclass: Introduce CACHED_CONFIGUREVARS
This variable is for holding the cached configure variables
to be specified in recipes. e.g.

CACHED_CONFIGUREVARS += "ac_cv_foo=yes ac_cv_bar=no"

This will make sure that the variables are not detected
by configure. This is useful in cross builds where
some features can not be detected correctly by configure
and having it as a variables gives us capability to
override it

Signed-off-by: Khem Raj <raj.khem@gmail.com>