alsa-state: move state files to localstatedir
alsactl creates the state files in /var/lib/alsa by default so switch
alsa-state to use files in that location.

Further, update the alsa-state init script to have the location of the
state files sed'ed into the script at do_install time (so as to remove
hard coding of directory paths).

Signed-off-by: Joshua Lock <josh@linux.intel.com>