pango: Use gnomebase class, not gnome
There was a circular dependency introduced by the recent gconf changes
to depend on gtk+. The issue is that gtk+ depends on pango and pango
depends on gconf.

This patch changes to use the gnonebase class since pango has no need
of gconf/mime/gtk-icon-cache and hence removes the circular dependency.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>