linux-yocto: add kernel 3.2.2 support
Adding initial support for the linux 3.2 kernel.

Although the branching and naming has changed slightly, the same BSPs and
features exist in this tree as have existed in previous linux-yocto versions.

Notable features of this update are:

  - streamlined branch naming
  - configuration and functionality split in the standard kernel
  - creation of the 'tiny' kernel type branch
  - audited kernel configurations with optional features documented
    in the meta branch
  - feature and BSP refreshes
  - 3.2-rt support refresh
  - qemu ppc graphics support

Build and boot testing has been performed on all emulated targets
(sato and core).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>