syslinux: Use SYSLINUX_TIMEOUT and SYSLINUX_PROMPT to configure syslinux
syslinux allows you to set TIMEOUT and PROMPT variables, when PROMPT is 0,
the "boot:" is not displayed uless one presses CTRL or SHIFT during startup.
TIMEOUT is in 1/10th of seconds, and a value of 0 for TIMEOUT will disable
the timeout mechanism.

In bitbake, recipes had set TIMEOUT (not SYSLINUX_TIMEOUT) incorrectly, other
patches fix this issues.  We are adding SYSLINUX_PROMPT to enable/disable the
"boot:" prompt in syslinux.

See http://www.syslinux.org/wiki/index.php/SYSLINUX for more details

Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Dexuan Cui <dexuan.cui@intel.com>