cross-canadian: Set STAGING_DIR_HOST correctly
As reported by Martin Jansa, the path to nativesdk sysroot was changing between
nativesdk and cross-canadian recipes. The problem was the incorrect deinfition of
STAGING_DIR_HOST in cross-canadian.bbclass.

Since nothing really uses the cross-canadian output in the sysroot, only the
packages, its not surprising this bug has gone un-noticed.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>