iptables: fix build error against 3.2+ kernel headers
The iptables local linux/types.h overrides the kernel/sysroot
types.h. As such, we need to provide some defines that are required
to build against 3.2+ kernel headers.

ifndef protection is provided for the defines to ensure that
configuration that already have these defines are still buildable.

This commit is temporary until a new version of iptables can be
used that contains the defines.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>