tiff: Make builds deterministic
libtiff now depends on lzma which can be obtained from xz and doesn't use lzo.
Previously, libtiff would detect and use lzma if it was present leading to
a number of race conditions including failures in things linking to libtiff
such as ghostscript since lzma could be removed while being rebuild leading
to failures in linking.

This patch corrects the dependency.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>